removed loginBox and registerBox, integrated them directly into login and register pages
